+++ packages/codegen/src/templates/processors/factory-builders.ts
@@ -5,7 +5,7 @@ export const factoryBuildersTemplate = ts`
  * WARNING: DO NOT EDIT
  * This file is auto-generated and updated by codegen
  */
-import type { ProcessorFactoryBuilder } from "@powerhousedao/reactor";
+import type { ProcessorFactoryBuilder } from "@powerhousedao/reactor-browser";
 
 export const processorFactoryBuilders: ProcessorFactoryBuilder[] = [];
 `.raw;